ISAR-Target-Classification/requirements.txt
numpy>=1.24.0
matplotlib>=3.7.0
pandas>=2.0.0
Pillow>=10.0.0
seaborn>=0.12.0
scikit-learn>=1.3.0
tensorflow>=2.14.0
